# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "AdminAction": {
        "description": "What an admin (or the command line, with no actor) did.",
        "enum": [
          "invite.create",
          "invite.revoke",
          "user.reset_link",
          "user.role_change",
          "user.deactivate",
          "user.reactivate",
          "user.delete"
        ],
        "title": "AdminAction",
        "type": "string"
      },
      "AdminEvent": {
        "description": "`actor` is null for the command line or a deleted admin, `target` for a deleted user.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AdminAction"
          },
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Details",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "actor",
          "action",
          "target",
          "details",
          "created_at"
        ],
        "title": "AdminEvent",
        "type": "object"
      },
      "AdminUser": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "is_active",
          "created_at",
          "last_seen_at"
        ],
        "title": "AdminUser",
        "type": "object"
      },
      "AdminUserUpdate": {
        "description": "Only the fields that are sent change.",
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "user",
                  "admin"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "title": "AdminUserUpdate",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "What went wrong; the frontend shows the translation `error.<code>`.",
        "enum": [
          "common.internal",
          "common.not_found",
          "common.method_not_allowed",
          "common.validation",
          "common.rate_limited",
          "common.unauthorized",
          "common.forbidden",
          "common.service_unavailable",
          "auth.invalid_credentials",
          "auth.account_deactivated",
          "auth.token_expired",
          "auth.session_expired",
          "auth.session_revoked",
          "auth.login_required",
          "auth.csrf",
          "auth.code_invalid",
          "auth.password_incorrect",
          "couple.already_in_couple",
          "couple.target_in_couple",
          "couple.request_pending",
          "admin.self_forbidden",
          "admin.last_admin",
          "admin.public_url_missing"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "description": "`params` fill placeholders in the translation; `fields` lists rejected request fields.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Fields",
            "type": "array"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code",
          "params",
          "fields"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "description": "One rejected request field; `loc` is its path, e.g. `[\"body\", \"name\"]`.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FieldErrorCode"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          }
        },
        "required": [
          "loc",
          "code"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FieldErrorCode": {
        "description": "Why a single request field was rejected (`fields[].code` in the envelope).",
        "enum": [
          "required",
          "invalid",
          "too_short",
          "too_long",
          "out_of_range",
          "invalid_format",
          "taken",
          "too_common",
          "same_as_username"
        ],
        "title": "FieldErrorCode",
        "type": "string"
      },
...
      "Invite": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "used",
              "expired",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          },
          "tailscale_share_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailscale Share Url"
          },
          "used_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "created_at",
          "expires_at",
          "created_by",
          "used_by",
          "tailscale_share_url"
        ],
        "title": "Invite",
        "type": "object"
      },
      "InviteCreate": {
        "description": "`tailscale_share_url`: the optional share link for step 1 of the invite message.",
        "properties": {
          "tailscale_share_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailscale Share Url"
          }
        },
        "title": "InviteCreate",
        "type": "object"
      },
      "InviteCreated": {
        "description": "The new invite and its link (`/join#<code>`); the code is shown only this once.",
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/Invite"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "invite",
          "url"
        ],
        "title": "InviteCreated",
        "type": "object"
      },
...
      "LinkCreated": {
        "description": "A reset link; the code is in the fragment (`/reset#<code>`).",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at"
        ],
        "title": "LinkCreated",
        "type": "object"
      },
...
      "UserRef": {
        "description": "Another user as everyone sees them; deactivated users are shown with a note (ADM-02).",
        "properties": {
          "deactivated": {
            "title": "Deactivated",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "deactivated"
        ],
        "title": "UserRef",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "The access token from login, refresh, join or reset.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "license": {
      "identifier": "AGPL-3.0-or-later",
      "name": "AGPL-3.0-or-later"
    },
    "title": "MealMate",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/events": {
      "get": {
        "description": "The admin activity log, newest first (at most 200 entries).",
        "operationId": "admin_list_events",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminEvent"
                  },
                  "title": "Response Admin List Events",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin List Events",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/invites": {
      "get": {
        "description": "All invites with their status, newest first.",
        "operationId": "admin_list_invites",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Invite"
                  },
                  "title": "Response Admin List Invites",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin List Invites",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "A new invite link (ACC-02); the link is returned only this once.",
        "operationId": "admin_create_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Create Invite",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/invites/{invite_id}": {
      "delete": {
        "description": "Revoke an invite so it can no longer be used.",
        "operationId": "admin_revoke_invite",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Revoke Invite",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "description": "All users, by display name.",
        "operationId": "admin_list_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUser"
                  },
                  "title": "Response Admin List Users",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "delete": {
        "description": "Delete another user and their data (ADM-03).",
        "operationId": "admin_delete_user",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Delete User",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "description": "Change another user's role or deactivate/reactivate them.",
        "operationId": "admin_update_user",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Update User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{user_id}/reset-link": {
      "post": {
        "description": "A single-use password reset link for the user (ACC-10); older ones stop working.",
        "operationId": "admin_create_reset_link",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Create Reset Link",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "description": "All active users except yourself, by display name.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRef"
                  },
                  "title": "Response List Users",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}